requirements.txt
pillow
python-dotenv
aiofiles
telethon